fix: replace simplified TJpgDec with official ChaN R0.03 source
- tjpgd.c/h: use official ChaN TJpgDec R0.03 from Bodmer/TJpg_Decoder
- Remove Bodmer-specific byte swap code (not needed for RGB888)
- Add tjpgdcnf.h configuration: RGB888, JD_FASTDECODE=1, JD_TBLCLIP=1
- jpeg_decoder.c: update input callback to size_t signature
- Root cause: simplified IDCT/Huffman implementation was producing
  corrupt output (dots on top, black on bottom)

+++ sle_mesh_networking/mesh/image/jpeg_decoder.c
@@ -47,10 +47,10 @@ typedef struct {
     uint32_t       pos;
 } mem_stream_t;
 
-static unsigned int jpeg_input_func(JDEC *jd, uint8_t *buff, unsigned int nbyte)
+static size_t jpeg_input_func(JDEC *jd, uint8_t *buff, size_t nbyte)
 {
     mem_stream_t *s = (mem_stream_t *)jd->device;
-    unsigned int remain = s->size - s->pos;
+    size_t remain = s->size - s->pos;
     if (nbyte > remain) nbyte = remain;
     if (buff) {
         (void)memcpy_s(buff, nbyte, s->data + s->pos, nbyte);
